Chunked resends are now sent "on-demand" instead of all-at-once
When MaxMessagesInResendRequest is set, QuickFIX generates multiple
"chunked" resends - and all of them at once. The CME rejects the second
resend complaining that a resend is already in progress (the first).

Modified the resend logic to send the next chunk only when the first one
is complete (i.e. when the last seqnum of the first has been received).