Chunked resends are now sent "on-demand" instead of all-at-once
When MaxMessagesInResendRequest is set, QuickFIX generates multiple
"chunked" resends - and all of them at once. The CME rejects the second
resend complaining that a resend is already in progress (the first).

Modified the resend logic to send the next chunk only when the first one
is complete (i.e. when the last seqnum of the first has been received).


Former-commit-id: d8671790d545936c0b3093550bfc31a1997196ab [formerly 3db17f76fd4cbe8e7b95ccbe20afba74167dcfaa] [formerly 861083739de132116f059daac5f3967a7b1b5e1d [formerly 889b8d8bdd70857337074bf1a46de22d32719f07]]
Former-commit-id: 8c1c89ac04531f2cc9b4a29122ca0e1ad9e483e0 [formerly d7fe448c36798c0f0692e9951afc0afd5872c9e0]
Former-commit-id: c576c3326e667db75ad6e64f56339486588d722d