Add new Application Callback (non-breaking) for early intercept of inbound messages.
Add an IApplicationExt interface that extends IApplication and facilitates early
intercept of inbound messages on the part of the Application.

This is a bit similar to the late outbound intercept afforded by ToApp/ToAdmin,
and permits alteration of field values prior to their being validated by the engine.
The difference is that, with the new interface, only one callback is provided for
both App and Admin messages.

Usage of the existing IApplication interface is unaffected.