Add new Application Callback (non-breaking) for early intercept of inbound messages.
Add an IApplicationExt interface that extends IApplication and facilitates early
intercept of inbound messages on the part of the Application.

This is a bit similar to the late outbound intercept afforded by ToApp/ToAdmin,
and permits alteration of field values prior to their being validated by the engine.
The difference is that, with the new interface, only one callback is provided for
both App and Admin messages.

Usage of the existing IApplication interface is unaffected.


Former-commit-id: df194d0fb5bb799e4aaa5eaae2e8d848cf7390af [formerly e239370631ac100aa4bda4b49dccb9660a368593] [formerly 54c09b79ac8d0b424a9a369a43c02fbc514ca6ba [formerly 4bc599bde671253f52348315967f316b44a7eb18]]
Former-commit-id: 65bc15c3919d08c2001f22dc7ceee73edc25a6c5 [formerly 1f53b6ba70a5b14b4ba5493ee8ef39aa584c2648]
Former-commit-id: 63440fd169a298984152c80a5a8397a92e10c8f9