Basic Session Lookup
Workaround method for when client messages and session config are
skweded due to clients sending in variations on what Qfn detects as a
Session.  For example, client sends in 57=ADMIN in session level
messages but not for business level messages.

GetBasicReverseSessionID() is used to find session on only Versio,
SenderCompID and TargetCompID and does not use SubID and LocationID
tags.