Updates project files for building NuGet packages
This commit updates the deliverable project files in preparation for
being able to build and push NuGet packages (and corresponding symbol
packages) via the package_release.bat batch script.

The copyright property was updated to show a proper date range which
auto-updates each year by using a computed value. The site URL, GitHub
repository URL, and license information were also added to properties in
the project files which is used by the packaging process.

Also updated the title property of the QuickFix project to be QuickFIX
Base to better clarify that this is the base (or core) library required
to implement an application using the FIX protocol and which does not
contain any of the protocol-specific messages.