Updates netcoreapp2.0 TFM to netcoreapp2.1
This commit updates the netcoreapp2.0 TFM to netcoreapp2.1 for the unit
tests project to remove a warning with respect to the NUnit Test
Adapter.

The AcceptanceTest project also used the net461 and
netcoreapp2.0 TFMs, so this was updated to also use the netcoreapp2.1
TFM as well. Lastly, the PowerShell script used to execute the
acceptance tests needed to be updated to use the dll with the correct
framework TFM.

The example projects also used the netcoreapp2.0 TFM and they
were updated, too.