Reformats acceptance_test.ps1
This commit largely reformats acceptance_test.ps1. But there were a few
modifications/additions.

Notably, instead of ALWAYS assuming that only the release configuration
would be run for the acceptance tests, you can now specify which build
configuration to use for running the acceptance tests. This can allow
easier debugging. If no configuration is specified, then by default, the
release configuration will be assumed.

And a UseWsl switch parameter is also added, which flows through the
various PowerShell scipts/funcitons which eventually allows one to use
their Ruby installation in their default WSL distribution, instead of
needing to have Ruby installed on their Windows machine. (Note to self:
should probably add an additional parameter that allows you to specify
WHICH WSL distro to use--in case you want to use a distro other than the
default.)

Another notable modification is that the main part of the script has
been wrapped in a try/finally block. It was possible before that if
there was an error running the acceptance tests, you'd be left in the
AcceptanceTests directory instead of back in the originating parent
directory. This ensures that if an error occurs, you are put back where
you started from.

And since the runat.ps1 script was changed to allow pipelined input, I
make use of that here to pass in the frameworks for which we're running
unit tests.