Fixes several issues with New-Release.ps1
The first issue fixed was that $BuildProps was being passed to dotnet as
a single string, and therefore, a single argument. So instead,
$BuildProps is now an array and is splatted to the call to dotnet
resulting in the desired behavior.

The second issue was found when running Ruby in WSL. Ruby expected the
paths to use '/' path separators--duh, it's Linux. A simple mistake on
my part.

When originally writing this script, I commented out the nuget push
stuff because I had nowhere to push packages--but I subsequently forgot
to uncomment that code out.

Lastly, I needed to fix up some paths when generating various artifacts.