Update SocketReader::ReadSome SocketError when socket closed
If 0 bytes are read it means the socet has been shutdown not reset.
Had this issue in production system reporting "Connection reset by peer" (the error message for SocketError.SocketReset), but using wireshark no RST packet was seen. However we did see the FIN and FIN,ACK packets which means the socket was shutdown. 
Setting to SocketError.Shutdown is the correct message for this scenario.