fix #963 NonSessionLog concurrency bug on Windows
Now each NSL.OnEvent call fully opens/writes/closes the log
in a sync block, so that two processes can never try to open it
at once.

OnEvent is only called rarely, so this should not pose
a performance issue.  (If OnEvent *is* called often enough
to be a perf issue, then you have a bigger problem than perf.)